+++ packages/@leight/file-server/src/endpoint/ChunkCommitEndpoint.ts
@@ -2,11 +2,9 @@ import {
     $ChunkService,
     type IChunkService,
     type IFileSourceSchema
-} from "@leight/file";
-import {
-    Endpoint,
-    type IEndpointFactory
-} from "@leight/next.js-server";
+}                              from "@leight/file";
+import {type IEndpointFactory} from "@leight/next.js";
+import {Endpoint}              from "@leight/next.js-server";
 
 /**
  * Export default this to handle chunk commits.