+++ packages/@leight/file-server/src/source/FileSource.ts
@@ -7,6 +7,10 @@ import {
     $PrismaClient,
     type PrismaClient
 }                       from "@leight/prisma";
+import {
+    type ISource,
+    withUpsert
+}                       from "@leight/source";
 import {AbstractSource} from "@leight/source-server";
 
 export class FileSource extends AbstractSource<IFileSourceSchema> implements IFileSource {
@@ -18,5 +22,13 @@ export class FileSource extends AbstractSource<IFileSourceSchema> implements IFi
         protected prismaClient: PrismaClient,
     ) {
         super($FileSource);
+    }
+
+    async runFind(id: string): Promise<IFileSourceSchema["Entity"]> {
+        return this.prismaClient.file.findUniqueOrThrow({where: {id}});
+    }
+
+    async runUpsert(props: ISource.IUpsert<IFileSourceSchema>): Promise<IFileSourceSchema["Entity"]> {
+        return this.prismaClient.file.upsert(withUpsert(props));
     }
 }