+++ packages/@leight/job-server/src/source/JobSource.ts
@@ -7,6 +7,7 @@ import {
     $PrismaClient,
     type PrismaClient
 }                       from "@leight/prisma";
+import {withPatch}      from "@leight/source";
 import {AbstractSource} from "@leight/source-server";
 
 export class JobSource extends AbstractSource<IJobSourceSchema> implements IJobSource {
@@ -18,5 +19,19 @@ export class JobSource extends AbstractSource<IJobSourceSchema> implements IJobS
         protected prismaClient: PrismaClient,
     ) {
         super($JobSource);
+    }
+
+    async runFind(id: string): Promise<IJobSourceSchema["Entity"]> {
+        return this.prismaClient.job.findUniqueOrThrow({where: {id}});
+    }
+
+    async runCreate(entity: IJobSourceSchema["Create"]): Promise<IJobSourceSchema["Entity"]> {
+        return this.prismaClient.job.create({
+            data: entity,
+        });
+    }
+
+    async runPatch(patch: IJobSourceSchema["Patch"]): Promise<IJobSourceSchema["Entity"]> {
+        return this.prismaClient.job.update(withPatch(patch));
     }
 }