+++ packages/@leight/next.js-server/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./api";
 export * from "./Endpoint";
 export * from "./NextAuthEndpoint";