+++ packages/@leight/filter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/filter
 
+## 0.5.30
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.29
 
 ### Patch Changes