+++ packages/@leight/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/i18n
 
+## 0.5.35
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.34
 
 ### Patch Changes