+++ packages/@leight/utils-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/utils-server
 
+## 0.5.34
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.33
 
 ### Patch Changes