+++ packages/@leight/xlsx-import-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @leight/xlsx-import-server
 
+## 0.5.34
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/job@0.5.38
+    - @leight/container@0.5.35
+    - @leight/file@0.5.35
+    - @leight/import@0.5.34
+    - @leight/trpc-server@0.5.37
+    - @leight/utils@0.5.37
+    - @leight/utils-server@0.5.34
+    - @leight/xlsx-import@0.5.35
+
 ## 0.5.33
 
 ### Patch Changes